# backend/postman/specs/openapi.yaml
openapi: 3.0.3
info:
  title: MSM — Sistema MIP de Productividad y Reconocimiento Comercial
  description: |
    API del sistema MSM (Sistema MIP) para medir, gestionar y reconocer el
    desempeño de la fuerza de ventas farmacéutica.

    **Autenticación**: Bearer JWT. Usar `POST /auth/login` con form-data
    (`username` + `password`) para obtener el `access_token`.

    **Roles**: ADMIN · PRESIDENCIA · DIR_COMERCIAL · GERENTE_PRODUCTIVIDAD ·
    GERENTE_DISTRITO · GERENTE_MARCA · REPRESENTANTE_MEDICO · CONSULTA
  version: 1.0.0
  contact:
    email: segura.moises@gmail.com
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /dashboard/comercial:
    get:
      tags: [Dashboard]
      summary: Dashboard Comercial (ventas, cuota, cumplimiento, crecimiento)
      parameters:
        - name: pais_id
          in: query
          schema: { type: integer }
        - name: ciclo_id
          in: query
          schema: { type: integer }
      responses:
        '200': { description: KPIs comerciales }
...
  /comercial/ventas:
    get:
      tags: [Comercial]
      summary: Datos de ventas paginados
      parameters:
        - name: pais_id
          in: query
          schema: { type: integer }
        - name: ciclo_id
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200': { description: Ventas paginadas }
  /comercial/evoir:
    get:
      tags: [Comercial]
      summary: Evolución de Prescripciones IR paginado
      parameters:
        - name: pais_id
          in: query
          schema: { type: integer }
        - name: ciclo_id
          in: query
          schema: { type: integer }
      responses:
        '200': { description: Datos EVOIR }
...